Refac: Updated folder structure and added tabs
